pyproject.toml
[project]
name = "ai-trading-agent"
version = "0.1.0"
description = "Multi-agent AI trading system with OpenAI integration"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "websockets>=11.0.0",
    "aiohttp>=3.8.0",
    "asyncio-mqtt>=0.16.0",
    "sqlalchemy>=2.0.0",
    "redis>=4.5.0",
    "psycopg2-binary>=2.9.0",
    "yfinance>=0.2.0",
    "alpha-vantage>=2.3.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "loguru>=0.7.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "schedule>=1.2.0",
    "ccxt>=4.0.0",
    "ta>=0.10.0",
    "plotly>=5.15.0",
    "dash>=2.14.0",
    # AI/LLM Dependencies
    "openai>=1.0.0",
    "anthropic>=0.8.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.0.20",
    "tiktoken>=0.5.0",
    # Multi-Agent Framework
    "autogen>=0.2.0",
    "crewai>=0.28.0",
    # Additional ML/AI tools
    "scikit-learn>=1.3.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "sentence-transformers>=2.2.0",
    # Async support
    "asyncio>=3.4.3",
    "aiofiles>=23.0.0",
    # Testing
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.poetry]
name = "ai-trading-agent"
version = "0.1.0"
description = "Multi-agent AI trading system with OpenAI integration"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
pandas = "^2.0.0"
numpy = "^1.24.0"
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"
loguru = "^0.7.0"
asyncio = "^3.4.3"
aiohttp = "^3.8.0"
sqlalchemy = "^2.0.0"
alembic = "^1.11.0"
psycopg2-binary = "^2.9.0"
redis = "^4.5.0"
celery = "^5.3.0"
fastapi = "^0.100.0"
uvicorn = "^0.23.0"
websockets = "^11.0.0"
ta = "^0.10.2"
yfinance = "^0.2.18"
alpha-vantage = "^2.3.1"
newsapi-python = "^0.2.7"
openai = "^1.0.0"
anthropic = "^0.7.0"
langchain = "^0.1.0"
langchain-openai = "^0.0.5"
autogen = "^0.2.0"
crewai = "^0.1.0"
streamlit = "^1.28.0"
plotly = "^5.17.0"
alpaca-trade-api = "^3.0.0"
ib-insync = "^0.9.86"
ccxt = "^4.1.0"
dash = "^2.14.0"
dash-bootstrap-components = "^1.5.0"
flask = "^3.0.0"
websocket-client = "^1.6.0"
schedule = "^1.2.0"
python-telegram-bot = "^20.6.0"
tweepy = "^4.14.0"
beautifulsoup4 = "^4.12.0"
selenium = "^4.15.0"
scikit-learn = "^1.3.0"
tensorflow = "^2.15.0"
torch = "^2.1.0"
transformers = "^4.35.0"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
jupyter = "^1.0.0"
ipykernel = "^6.26.0"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.9.0"
flake8 = "^6.1.0"
mypy = "^1.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.9.0"
flake8 = "^6.1.0"
mypy = "^1.6.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = [
    "tests",
]
python_files = [
    "test_*.py",
    "*_test.py",
]
asyncio_mode = "auto"
